Extract query building
